+++ system/Drivers/STM32G4xx_HAL_Driver/Inc/stm32g4xx_ll_cordic.h
@@ -676,12 +676,12 @@ __STATIC_INLINE uint32_t LL_CORDIC_DMA_GetRegAddr(CORDIC_TypeDef *CORDICx, uint3
   if (Direction == LL_CORDIC_DMA_REG_DATA_OUT)
   {
     /* return address of RDATA register */
-    data_reg_addr = (uint32_t) & (CORDICx->RDATA);
+    data_reg_addr = (uintptr_t) & (CORDICx->RDATA);
   }
   else
   {
     /* return address of WDATA register */
-    data_reg_addr = (uint32_t) & (CORDICx->WDATA);
+    data_reg_addr = (uintptr_t) & (CORDICx->WDATA);
   }
 
   return data_reg_addr;
