+++ system/Drivers/STM32G4xx_HAL_Driver/Inc/stm32g4xx_ll_opamp.h
@@ -73,7 +73,7 @@ extern "C" {
   * @retval Register address
   */
 #define __OPAMP_PTR_REG_OFFSET(__REG__, __REG_OFFSET__)                        \
-  ((__IO uint32_t *)((uint32_t) ((uint32_t)(&(__REG__)) + ((__REG_OFFSET__) << 2))))
+  ((__IO uint32_t *)((uint32_t) ((uintptr_t)(&(__REG__)) + ((__REG_OFFSET__) << 2))))
