+++ system/Drivers/STM32L4xx_HAL_Driver/Inc/stm32l4xx_ll_opamp.h
@@ -86,7 +86,7 @@ extern "C" {
   * @retval Register address
 */
 #define __OPAMP_PTR_REG_OFFSET(__REG__, __REG_OFFSET__)                        \
- ((uint32_t *)((uint32_t) ((uint32_t)(&(__REG__)) + ((__REG_OFFSET__) << 2U))))
+ ((uint32_t *)((uint32_t) ((uintptr_t)(&(__REG__)) + ((__REG_OFFSET__) << 2U))))
