+++ system/Drivers/STM32L5xx_HAL_Driver/Inc/stm32l5xx_ll_lpuart.h
@@ -2490,12 +2490,12 @@ __STATIC_INLINE uint32_t LL_LPUART_DMA_GetRegAddr(USART_TypeDef *LPUARTx, uint32
   if (Direction == LL_LPUART_DMA_REG_DATA_TRANSMIT)
   {
     /* return address of TDR register */
-    data_reg_addr = (uint32_t) &(LPUARTx->TDR);
+    data_reg_addr = (uintptr_t) &(LPUARTx->TDR);
   }
   else
   {
     /* return address of RDR register */
-    data_reg_addr = (uint32_t) &(LPUARTx->RDR);
+    data_reg_addr = (uintptr_t) &(LPUARTx->RDR);
   }
 
   return data_reg_addr;
